requirements.dev.txt
Django>=5.2.1,<5.3
djangorestframework>=3.16,<3.17
psycopg2
python-decouple>=3.8,<3.9
pillow>=11.2.1,<11.3
django-filter>=25.1,<25.2


